+++ glean-core/tests/common/mod.rs
@@ -94,7 +94,7 @@ pub fn get_queued_pings(data_path: &Path) -> Result<Vec<(String, JsonValue)>> {
     get_pings(&data_path.join("pending_pings"))
 }
 
-/// Get a vector of the currently queued `deletion_request` pings.
+/// Get a vector of the currently queued `deletion-request` pings.
 ///
 /// # Arguments
 ///
@@ -102,7 +102,7 @@ pub fn get_queued_pings(data_path: &Path) -> Result<Vec<(String, JsonValue)>> {
 ///
 /// # Returns
 ///
-/// A vector of all queued `deletion_request` pings. Each entry is a pair `(url, json_data)`,
+/// A vector of all queued `deletion-request` pings. Each entry is a pair `(url, json_data)`,
 /// where `url` is the endpoint the ping will go to, and `json_data` is the JSON payload.
 pub fn get_deletion_pings(data_path: &Path) -> Result<Vec<(String, JsonValue)>> {
     get_pings(&data_path.join("deletion_request"))